+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Errors are now printed to `stderr` instead of `stdout`
 - Date higlighter now supports different highlights for date and time segment
+- Added new Key Value highlighter
 
 
 ## 1.4.0 - 12.08.23